+++ rust/src/router/data_consumer.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod tests;
 
-use crate::data_producer::{DataProducer, DataProducerId};
+use crate::data_producer::{DataProducer, DataProducerId, WeakDataProducer};
 use crate::data_structures::{AppData, WebRtcMessage};
 use crate::messages::{
     DataConsumerCloseRequest, DataConsumerDumpRequest, DataConsumerGetBufferedAmountRequest,
@@ -198,7 +198,7 @@ struct Inner {
     sctp_stream_parameters: Option<SctpStreamParameters>,
     label: String,
     protocol: String,
-    data_producer: DataProducer,
+    data_producer_id: DataProducerId,
     direct: bool,
     executor: Arc<Executor<'static>>,
     channel: Channel,
@@ -206,7 +206,8 @@ struct Inner {
     handlers: Arc<Handlers>,
     app_data: AppData,
     transport: Box<dyn Transport>,
-    closed: AtomicBool,
+    weak_data_producer: WeakDataProducer,
+    closed: Arc<AtomicBool>,
     // Drop subscription to consumer-specific notifications when consumer itself is dropped
     _subscription_handlers: Vec<Option<SubscriptionHandler>>,
     _on_transport_close_handler: Mutex<HandlerId>,
@@ -234,16 +235,18 @@ impl Inner {
                         router_id: self.transport.router_id(),
                         transport_id: self.transport.id(),
                         data_consumer_id: self.id,
-                        data_producer_id: self.data_producer.id(),
+                        data_producer_id: self.data_producer_id,
                     },
                 };
-                let data_producer = self.data_producer.clone();
+                let data_producer = self.weak_data_producer.upgrade();
                 let transport = self.transport.clone();
 
                 self.executor
                     .spawn(async move {
-                        if let Err(error) = channel.request(request).await {
-                            error!("consumer closing failed on drop: {}", error);
+                        if data_producer.is_some() {
+                            if let Err(error) = channel.request(request).await {
+                                error!("consumer closing failed on drop: {}", error);
+                            }
                         }
 
                         drop(data_producer);
@@ -271,7 +274,7 @@ impl fmt::Debug for RegularDataConsumer {
             .field("sctp_stream_parameters", &self.inner.sctp_stream_parameters)
             .field("label", &self.inner.label)
             .field("protocol", &self.inner.protocol)
-            .field("data_producer_id", &self.inner.data_producer.id())
+            .field("data_producer_id", &self.inner.data_producer_id)
             .field("transport", &self.inner.transport)
             .field("closed", &self.inner.closed)
             .finish()
@@ -299,7 +302,7 @@ impl fmt::Debug for DirectDataConsumer {
             .field("sctp_stream_parameters", &self.inner.sctp_stream_parameters)
             .field("label", &self.inner.label)
             .field("protocol", &self.inner.protocol)
-            .field("data_producer_id", &self.inner.data_producer.id())
+            .field("data_producer_id", &self.inner.data_producer_id)
             .field("transport", &self.inner.transport)
             .field("closed", &self.inner.closed)
             .finish()
@@ -358,20 +361,25 @@ impl DataConsumer {
         debug!("new()");
 
         let handlers = Arc::<Handlers>::default();
+        let closed = Arc::new(AtomicBool::new(false));
 
         let inner_weak = Arc::<Mutex<Option<Weak<Inner>>>>::default();
         let subscription_handler = {
             let handlers = Arc::clone(&handlers);
+            let closed = Arc::clone(&closed);
             let inner_weak = Arc::clone(&inner_weak);
 
             channel.subscribe_to_notifications(id.into(), move |notification| {
                 match serde_json::from_value::<Notification>(notification) {
                     Ok(notification) => match notification {
                         Notification::DataProducerClose => {
-                            handlers.data_producer_close.call_simple();
-                            if let Some(inner) = inner_weak.lock().as_ref().and_then(Weak::upgrade)
-                            {
-                                inner.close(false);
+                            if !closed.load(Ordering::SeqCst) {
+                                handlers.data_producer_close.call_simple();
+                                if let Some(inner) =
+                                    inner_weak.lock().as_ref().and_then(Weak::upgrade)
+                                {
+                                    inner.close(false);
+                                }
                             }
                         }
                         Notification::SctpSendBufferFull => {
@@ -433,7 +441,7 @@ impl DataConsumer {
             sctp_stream_parameters,
             label,
             protocol,
-            data_producer,
+            data_producer_id: data_producer.id(),
             direct,
             executor,
             channel,
@@ -441,7 +449,8 @@ impl DataConsumer {
             handlers,
             app_data,
             transport,
-            closed: AtomicBool::new(false),
+            weak_data_producer: data_producer.downgrade(),
+            closed,
             _subscription_handlers: vec![subscription_handler, payload_subscription_handler],
             _on_transport_close_handler: Mutex::new(on_transport_close_handler),
         });
@@ -464,7 +473,7 @@ impl DataConsumer {
     /// The associated data producer identifier.
     #[must_use]
     pub fn data_producer_id(&self) -> DataProducerId {
-        self.inner().data_producer.id()
+        self.inner().data_producer_id
     }
 
     /// The type of the data consumer.
@@ -659,7 +668,7 @@ impl DataConsumer {
             router_id: self.inner().transport.router_id(),
             transport_id: self.inner().transport.id(),
             data_consumer_id: self.inner().id,
-            data_producer_id: self.inner().data_producer.id(),
+            data_producer_id: self.inner().data_producer_id,
         }
     }
 }
@@ -677,7 +686,7 @@ impl DirectDataConsumer {
                         router_id: self.inner.transport.router_id(),
                         transport_id: self.inner.transport.id(),
                         data_consumer_id: self.inner.id,
-                        data_producer_id: self.inner.data_producer.id(),
+                        data_producer_id: self.inner.data_producer_id,
                     },
                     data: DataConsumerSendRequestData { ppid },
                 },